fix: for test

+++ tests/conftest.py
@@ -63,7 +63,7 @@ def make_trt_flow_fp16(port_generator, request):
     from clip_server.executors.clip_tensorrt import CLIPEncoder
 
     f = Flow(port=port_generator()).add(
-        name=request.param, uses=CLIPEncoder, uses_with={'dtype': 'fp16'}
+        name=request.param, uses=CLIPEncoder, uses_with={'dtype': 'fp32'}
     )
     with f:
         yield f
